requirements.txt
# Data Processing
pandas==2.1.0
numpy==1.24.3

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1

# API
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.5.2

# NLP & ML
transformers==4.34.0
torch==2.1.0
scikit-learn==1.3.0
konlpy==0.6.0

# AWS
boto3==1.28.36

# Kafka
kafka-python==2.0.2

# Database
sqlalchemy==2.0.21
pymongo==4.5.0
motor==3.3.2  # MongoDB 비동기 드라이버

# Testing
pytest==7.4.2
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.17.1

# Development
ruff==0.1.8
mypy==1.7.1

# Dashboard
streamlit
tenacity
